PR feat-web-app (#6)
* fix(docs): update file references in mkdocs and deployment workflow

* fix(docu): update version badge color in README and bumpversion configuration

* refactor(workflow): simplify version bumping process by removing commit and tag steps

* fix(workflow): move Git setup to a separate step in bump version workflow

* fix(workflow): add verbose output to version bumping step in bump version workflow

* fix(workflow): comment out version bump command for debugging purposes

* fix(workflow): comment out version bump command for debugging

* fix(workflow): enhance installation process for pandoc and texlive tools

* fix(workflow): add documentation comments to Makefile targets for clarity

* fix(workflow): uncomment version bump command to enable versioning

* fix(workflow): refactor bump version workflow for improved clarity and functionality

* fix(workflow): update bump version workflow and add scripts for PR creation and branch deletion

* fix(workflow): update actions/checkout and bump-my-version versions for consistency

* fix(workflow): update conditional check for clarity and linting

* fix(workflow): update lychee-action and create-issue-from-file versions for consistency

+++ Makefile
@@ -13,7 +13,7 @@ VENV_PATH := $(ROOT_PATH)/.venv
 LOG_PATH := $(ROOT_PATH)/logs
 AZD_URL := https://aka.ms/install-azd.sh
 
-setup:
+setup:  ## Sets up the tool chain and installs dependencies
 	echo "Setting up tool chain ..."
 	cat /etc/os-release | grep PRETTY_NAME
 	pip3 install --no-cache-dir --disable-pip-version-check uv
@@ -21,7 +21,7 @@ setup:
 	$(MAKE) -s download_azd
 	$(MAKE) -s install_pandoc_tex
 
-download_azd:
+download_azd:  ## Downloads and installs azd (Azure Developer CLI)
 	echo "Setting up azd ..."
 	if ! which "azd" > /dev/null 2>&1; then
 		curl -fsSL "$(AZD_URL)" | bash
@@ -29,47 +29,69 @@ download_azd:
 		echo "azd already present."
 	fi
 
-install_pandoc_tex:
+install_pandoc_tex:  ## Installs pandoc and texlive
 	echo "Setting up pandoc and texlive ..."
-	if [ -n "$$(sudo --version 2>/dev/null)" ]; then
-		SUDO="sudo"
-	else
-		SUDO=""
+	# tools_pkgs="pandoc texlive"
+	tools_cmds="pandoc pdflatex"
+	tools_to_install=""
+	for tool in $$tools_cmds; do
+		if $$tool --version >/dev/null 2>&1; then
+			echo "$${tool} already installed: $$($$tool --version | head -n 1)"
+		else
+			tools_to_install="$${tools_to_install} $${tool}"
+		fi
+	done
+	if [ -n "$$tools_to_install" ]; then
+		# check if sudo present, assume root otherwise
+		if [ -n "$$(sudo --version 2>/dev/null)" ]; then
+			SUDO="sudo"
+		else
+			SUDO=""
+		fi
+		$$SUDO apt-get update -qq
+		# known good: pandoc 2.17.1.1, pdfTeX 3.141592653-2.6-1.40.24
+		for tool in $$tools_to_install; do
+			# TODO  remove case for indexing $tools_pkgs with $tools_cmds
+			case "$$tool" in
+				pandoc) tool_to_install="pandoc" ;;
+				pdflatex) tool_to_install="texlive" ;;
+				*) echo "Unknown tool. Exiting iteration ..." && continue ;;
+			esac
+			echo "Installing $${tool_to_install} ..."
+			$$SUDO apt-get install $$tool_to_install -y -qq 1>/dev/null || \
+				{ echo "Failed to install $${tool_to_install}. Exiting ..."; continue; }
+			$$tool --version | head -n 1
+		done
+		$$SUDO apt-get install -f -y -qq
+		$$SUDO apt-get autoclean -y -qq
+		$$SUDO apt-get autoremove -y -qq
 	fi
-	$$SUDO apt-get update -qq
-	# known good: pandoc 2.17.1.1, pdfTeX 3.141592653-2.6-1.40.24
-	$$SUDO apt-get install pandoc texlive -y -qq
-	$$SUDO apt-get install -f -y -qq
-	$$SUDO apt-get autoclean -y -qq
-	$$SUDO apt-get autoremove -y -qq
-	pandoc --version | head -n 1
-	pdflatex --version | head -n 1
 
-run_local:
+run_local:  ## Runs the app locally with uvicorn
 	$(MAKE) -s ruff
 	mkdir -p "$(LOG_PATH)"
 	SYS_ROOT_PATH="$(ROOT_PATH)" uv run uvicorn $(APP_START) --reload
 
-build_local:
+build_local:  ## Builds the app locally with uv
 	$(MAKE) -s ruff
 	$(MAKE) -s export_reqs
 	chmod +x "$(STARTUP_FILE)"
 	SYS_ROOT_PATH="$(ROOT_PATH)" VENV_PATH="$(VENV_PATH)" bash -x "$(STARTUP_FILE)"
 
-ruff:
+ruff:  ## Runs ruff to check and format the code
 	uv run ruff check --fix
 	uv run ruff format
 
-export_reqs:
+export_reqs:  ## Exports pyproject requirements to requirements.txt
 	uv export --format requirements-txt --no-dev > "$(REQS_FILE)"
 
-test_all:
+test_all:  ## Runs all tests with pytest
 	SYS_ROOT_PATH="$(ROOT_PATH)" uv run pytest --tb=short
 
-type_check:
+type_check:  ## Runs mypy for type checking
 	uv run mypy src
 
-bump_dry:  # Runs bump-my-version strat=[strat] in dry-mode to show what-if
+bump_dry:  ## Runs bump-my-version strat=[strat] in dry-mode to show what-if
 	if [ -z "$$strat" ];
 		then strat="minor";
 	fi
